refactor/docs: split up long modules and add some missing documentation
