+++ backend/api/src/db/image.rs
@@ -9,299 +9,9 @@ use shared::domain::{
 use sqlx::{PgConnection, PgPool};
 use uuid::Uuid;
 
-pub mod user {
-    use futures::stream::BoxStream;
-    use shared::domain::image::{user::UserImage, ImageId};
-    use sqlx::PgPool;
-
-    pub async fn create(conn: &PgPool) -> sqlx::Result<ImageId> {
-        let id: ImageId = sqlx::query!(
-            r#"
-insert into user_image_library default values
-returning id as "id: ImageId"
-"#,
-        )
-        .fetch_one(conn)
-        .await?
-        .id;
-
-        Ok(id)
-    }
-
-    pub async fn delete(db: &PgPool, image: ImageId) -> sqlx::Result<()> {
-        sqlx::query!("delete from user_image_library where id = $1", image.0)
-            .execute(db)
-            .await
-            .map(drop)
-    }
-
-    pub async fn get(db: &PgPool, image: ImageId) -> sqlx::Result<Option<UserImage>> {
-        sqlx::query_as!(
-            UserImage,
-            r#"select id as "id: ImageId" from user_image_library where id = $1"#,
-            image.0
-        )
-        .fetch_optional(db)
-        .await
-    }
-
-    pub fn list(db: &PgPool) -> BoxStream<'_, sqlx::Result<UserImage>> {
-        sqlx::query_as!(
-            UserImage,
-            r#"select id as "id: ImageId" from user_image_library order by created_at desc"#,
-        )
-        .fetch(db)
-    }
-}
-
-pub mod tag {
-    use crate::error;
-    use paperclip::actix::NoContent;
-    use shared::domain::{image::tag::ImageTagResponse, meta::TagId};
-    use sqlx::postgres::PgDatabaseError;
-    use sqlx::PgPool;
-
-    pub async fn list(db: &PgPool) -> sqlx::Result<Vec<ImageTagResponse>> {
-        sqlx::query_as!(
-            ImageTagResponse,
-            r#"
-select id as "id: TagId", display_name, index from "image_tag"
-order by index
-            "#
-        )
-        .fetch_all(db)
-        .await
-    }
-
-    pub async fn create(
-        db: &PgPool,
-        index: i16,
-        display_name: &str,
-    ) -> Result<(i16, String, TagId), error::Tag> {
-        let mut txn = db.begin().await?;
-
-        let res = sqlx::query!(
-            // language=SQL
-            r#"
-insert into image_tag (index, display_name)
-values ($1, $2)
-returning id as "id: TagId", index as "index: i16", display_name
-            "#,
-            index,
-            display_name,
-        )
-        .fetch_one(&mut txn)
-        .await?;
-
-        txn.commit().await.map_err(|err| match err {
-            sqlx::Error::Database(err)
-                if err.downcast_ref::<PgDatabaseError>().constraint()
-                    == Some("image_tag_index_key") =>
-            {
-                log::info!("CONFLICT FOUND");
-                error::Tag::TakenIndex
-            }
-            e => e.into(),
-        })?;
-
-        Ok((res.index, res.display_name, res.id))
-    }
-
-    pub async fn update(
-        db: &PgPool,
-        curr_index: i16,
-        display_name: Option<&str>,
-        new_index: Option<i16>,
-    ) -> Result<NoContent, error::Tag> {
-        let mut txn = db.begin().await?;
-
-        let res = sqlx::query!(
-            r#"select index as "index: i16" from image_tag where index = $1 for update"#,
-            curr_index
-        )
-        .fetch_optional(&mut txn)
-        .await?;
-
-        if res.is_none() {
-            txn.commit().await?;
-            return Err(error::Tag::ResourceNotFound);
-        }
-
-        if let Some(display_name) = display_name {
-            sqlx::query!(
-                r#"update image_tag set display_name = $2 where index = $1"#,
-                curr_index,
-                display_name,
-            )
-            .execute(&mut txn)
-            .await?;
-        }
-
-        if let Some(new_index) = new_index {
-            sqlx::query!(
-                r#"
-            update image_tag set index = $2 where index = $1
-            "#,
-                curr_index,
-                new_index
-            )
-            .execute(&mut txn)
-            .await?;
-        }
-
-        txn.commit().await.map_err(|err| match err {
-            sqlx::Error::Database(err)
-                if err.downcast_ref::<PgDatabaseError>().constraint()
-                    == Some("image_tag_index_key") =>
-            {
-                error::Tag::TakenIndex
-            }
-            e => e.into(),
-        })?;
-
-        Ok(NoContent)
-    }
-
-    pub async fn delete(db: &PgPool, index: i16) -> sqlx::Result<()> {
-        sqlx::query!(
-            r#"
-delete from image_tag where index = $1
-            "#,
-            index
-        )
-        .execute(db)
-        .await?;
-
-        Ok(())
-    }
-}
-
-pub mod recent {
-    use crate::error;
-    use chrono::{DateTime, Utc};
-    use shared::{
-        domain::image::{recent::UserRecentImageResponse, ImageId},
-        media::MediaLibrary,
-    };
-    use sqlx::PgPool;
-    use uuid::Uuid;
-
-    pub async fn create(
-        db: &PgPool,
-        user_id: Uuid,
-        image_id: ImageId,
-        library: MediaLibrary,
-    ) -> anyhow::Result<(ImageId, MediaLibrary, DateTime<Utc>), error::UserRecentImage> {
-        let mut txn = db.begin().await?;
-
-        let exists = sqlx::query!(
-            r#"
-select exists(select 1 from user_recent_image where user_id = $1 and image_id = $2) as "exists!"
-            "#,
-            user_id,
-            image_id.0,
-        )
-        .fetch_one(&mut txn)
-        .await?
-        .exists;
-
-        if exists {
-            return Err(error::UserRecentImage::Conflict);
-        }
-
-        let res = sqlx::query!(
-            r#"
-insert into user_recent_image (user_id, image_id, media_library)
-values ($1, $2, $3)
-returning image_id as "id: ImageId", media_library as "library: MediaLibrary", last_used as "last_used: DateTime<Utc>";
-            "#,
-            user_id,
-            image_id.0,
-            library as i16,
-        )
-        .fetch_one(&mut txn)
-        .await?;
-
-        txn.commit().await?;
-
-        Ok((res.id, res.library, res.last_used))
-    }
-
-    pub async fn update(
-        db: &PgPool,
-        user_id: Uuid,
-        image_id: ImageId,
-    ) -> anyhow::Result<(), error::UserRecentImage> {
-        let mut txn = db.begin().await?;
-
-        let exists = sqlx::query!(
-            r#"
-select exists(select 1 from user_recent_image where user_id = $1 and image_id = $2) as "exists!"
-            "#,
-            user_id,
-            image_id.0,
-        )
-        .fetch_one(&mut txn)
-        .await?
-        .exists;
-
-        if !exists {
-            return Err(error::UserRecentImage::ResourceNotFound);
-        }
-
-        sqlx::query!(
-            r#"
-update user_recent_image
-set last_used = now()
-where user_id = $1 and image_id = $2
-            "#,
-            user_id,
-            image_id.0,
-        )
-        .execute(&mut txn)
-        .await?;
-
-        txn.commit().await.map_err(Into::into)
-    }
-
-    pub async fn delete(db: &PgPool, user_id: Uuid, image_id: ImageId) -> sqlx::Result<()> {
-        sqlx::query!(
-            r#"
-delete from user_recent_image
-where user_id = $1 and image_id = $2
-            "#,
-            user_id,
-            image_id.0
-        )
-        .execute(db)
-        .await?;
-
-        Ok(())
-    }
-
-    pub async fn list(
-        db: &PgPool,
-        user_id: Uuid,
-        limit: Option<i64>,
-    ) -> sqlx::Result<Vec<UserRecentImageResponse>> {
-        // if let Some(limit) = limit { assert!(limit >= 0); }
-
-        sqlx::query_as!(
-            UserRecentImageResponse,
-            r#"
-select image_id as "id: ImageId", media_library as "library: MediaLibrary", last_used as "last_used: DateTime<Utc>"
-from user_recent_image
-where user_id = $1
-order by last_used desc
-limit $2
-            "#,
-            user_id,
-            limit
-        )
-        .fetch_all(db)
-        .await
-    }
-}
+pub mod recent;
+pub mod tag;
+pub mod user;
 
 pub async fn create(
     conn: &mut PgConnection,
