+++ backend/api/src/http/endpoints/image.rs
@@ -23,302 +23,9 @@ use shared::{
 };
 use sqlx::{postgres::PgDatabaseError, PgPool};
 
-pub mod user {
-    use crate::{
-        db, error, extractor::TokenUser, image_ops::generate_images, s3, service::ServiceData,
-    };
-    use paperclip::actix::{
-        api_v2_operation,
-        web::{Bytes, Data, Json, Path},
-        CreatedJson, NoContent,
-    };
-
-    use futures::TryStreamExt;
-    use shared::{
-        api::{endpoints, ApiEndpoint},
-        domain::{
-            image::{
-                user::{UserImage, UserImageListResponse, UserImageResponse},
-                ImageId, ImageKind,
-            },
-            CreateResponse,
-        },
-        media::MediaLibrary,
-        media::{FileKind, PngImageFile},
-    };
-    use sqlx::PgPool;
-
-    /// Create a image in the user's image library.
-    #[api_v2_operation]
-    pub(super) async fn create(
-        db: Data<PgPool>,
-        _claims: TokenUser,
-    ) -> Result<CreatedJson<<endpoints::image::user::Create as ApiEndpoint>::Res>, error::Server>
-    {
-        let id = db::image::user::create(db.as_ref()).await?;
-        Ok(CreatedJson(CreateResponse { id }))
-    }
-
-    /// Upload an image to the user's image library.
-    #[api_v2_operation]
-    pub(super) async fn upload(
-        db: Data<PgPool>,
-        s3: ServiceData<s3::Client>,
-        _claims: TokenUser,
-        Path(id): Path<ImageId>,
-        bytes: Bytes,
-    ) -> Result<NoContent, error::Upload> {
-        let mut txn = db.begin().await?;
-
-        sqlx::query!(
-            r#"select 1 as discard from user_image_library where id = $1 for update"#,
-            id.0
-        )
-        .fetch_optional(&mut txn)
-        .await?
-        .ok_or(error::Upload::ResourceNotFound)?;
-
-        let kind = ImageKind::Sticker;
-
-        let (original, resized, thumbnail) =
-            actix_web::web::block(move || -> Result<_, error::Upload> {
-                let original =
-                    image::load_from_memory(&bytes).map_err(|_| error::Upload::InvalidMedia)?;
-                Ok(generate_images(&original, kind)?)
-            })
-            .await
-            .map_err(error::Upload::blocking_error)?;
-
-        s3.upload_png_images(MediaLibrary::User, id.0, original, resized, thumbnail)
-            .await?;
-
-        sqlx::query!(
-            "update user_image_library set uploaded_at = now() where id = $1",
-            id.0
-        )
-        .execute(&mut txn)
-        .await?;
-
-        txn.commit().await?;
-
-        Ok(NoContent)
-    }
-
-    /// Delete an image from the user's image library.
-    #[api_v2_operation]
-    pub(super) async fn delete(
-        db: Data<PgPool>,
-        _claims: TokenUser,
-        req: Path<ImageId>,
-        s3: ServiceData<s3::Client>,
-    ) -> Result<NoContent, error::Delete> {
-        let image = req.into_inner();
-        db::image::user::delete(&db, image)
-            .await
-            .map_err(super::check_conflict_delete)?;
-
-        let delete = |kind| s3.delete_media(MediaLibrary::User, FileKind::ImagePng(kind), image.0);
-        let ((), (), ()) = futures::future::join3(
-            delete(PngImageFile::Original),
-            delete(PngImageFile::Resized),
-            delete(PngImageFile::Thumbnail),
-        )
-        .await;
-
-        Ok(NoContent)
-    }
-
-    /// Get an image from the user's image library.
-    #[api_v2_operation]
-    pub(super) async fn get(
-        db: Data<PgPool>,
-        _claims: TokenUser,
-        req: Path<ImageId>,
-    ) -> Result<Json<<endpoints::image::user::Get as ApiEndpoint>::Res>, error::NotFound> {
-        let metadata = db::image::user::get(&db, req.into_inner())
-            .await?
-            .ok_or(error::NotFound::ResourceNotFound)?;
-
-        Ok(Json(UserImageResponse { metadata }))
-    }
-
-    /// List images from the user's image library.
-    #[api_v2_operation]
-    pub(super) async fn list(
-        db: Data<PgPool>,
-        _claims: TokenUser,
-    ) -> Result<Json<<endpoints::image::user::List as ApiEndpoint>::Res>, error::Server> {
-        let images: Vec<_> = db::image::user::list(db.as_ref())
-            .err_into::<error::Server>()
-            .and_then(|metadata: UserImage| async { Ok(UserImageResponse { metadata }) })
-            .try_collect()
-            .await?;
-
-        Ok(Json(UserImageListResponse { images }))
-    }
-}
-
-/// Manage image tags
-pub mod tag {
-    use crate::{
-        db, error,
-        extractor::{ScopeAdmin, TokenUserWithScope},
-    };
-    use paperclip::actix::web::Path;
-    use paperclip::actix::{
-        api_v2_operation,
-        web::{Data, Json},
-        CreatedJson, NoContent,
-    };
-    use shared::domain::image::tag::ImageTagResponse;
-    use shared::{
-        api::{endpoints, ApiEndpoint},
-        domain::image::tag::ImageTagListResponse,
-    };
-    use sqlx::PgPool;
-
-    #[api_v2_operation]
-    pub(super) async fn list(
-        db: Data<PgPool>,
-        _claims: TokenUserWithScope<ScopeAdmin>,
-    ) -> Result<Json<<endpoints::image::tag::List as ApiEndpoint>::Res>, error::NotFound> {
-        log::info!("reached List");
-
-        let image_tags = db::image::tag::list(db.as_ref()).await?;
-
-        Ok(Json(ImageTagListResponse { image_tags }))
-    }
-
-    #[api_v2_operation]
-    pub(super) async fn create(
-        db: Data<PgPool>,
-        _claims: TokenUserWithScope<ScopeAdmin>,
-        index: Path<i16>,
-        req: Json<<endpoints::image::tag::Create as ApiEndpoint>::Req>,
-    ) -> Result<CreatedJson<<endpoints::image::tag::Create as ApiEndpoint>::Res>, error::Tag> {
-        let res =
-            db::image::tag::create(db.as_ref(), index.into_inner(), req.display_name.as_str())
-                .await?;
-
-        Ok(CreatedJson(ImageTagResponse {
-            index: res.0,
-            display_name: res.1,
-            id: res.2,
-        }))
-    }
-
-    #[api_v2_operation]
-    pub(super) async fn update(
-        db: Data<PgPool>,
-        _claims: TokenUserWithScope<ScopeAdmin>,
-        index: Path<i16>,
-        req: Json<<endpoints::image::tag::Update as ApiEndpoint>::Req>,
-    ) -> Result<NoContent, error::Tag> {
-        let req = req.into_inner();
-
-        let _resp = db::image::tag::update(
-            db.as_ref(),
-            index.into_inner(),
-            req.display_name.as_deref(),
-            req.index,
-        )
-        .await?;
-
-        Ok(NoContent)
-    }
-
-    #[api_v2_operation]
-    pub(super) async fn delete(
-        db: Data<PgPool>,
-        _claims: TokenUserWithScope<ScopeAdmin>,
-        req: Path<i16>,
-    ) -> Result<NoContent, error::Tag> {
-        db::image::tag::delete(db.as_ref(), req.into_inner()).await?;
-
-        Ok(NoContent)
-    }
-}
-
-pub mod recent {
-    use crate::{db, error, extractor::TokenUser};
-    use chrono::{DateTime, Utc};
-    use paperclip::actix::{
-        api_v2_operation,
-        web::{Data, Json, Path, Query},
-        CreatedJson, NoContent,
-    };
-    use shared::{
-        api::{endpoints::image::recent, ApiEndpoint},
-        domain::image::{
-            recent::{UserRecentImageListResponse, UserRecentImageResponse},
-            ImageId,
-        },
-        media::MediaLibrary,
-    };
-    use sqlx::PgPool;
-
-    #[api_v2_operation]
-    pub(in super::super) async fn create(
-        db: Data<PgPool>,
-        claims: TokenUser,
-        req: Json<<recent::Create as ApiEndpoint>::Req>,
-    ) -> Result<CreatedJson<UserRecentImageResponse>, error::UserRecentImage> {
-        let req = req.into_inner();
-
-        let (id, library, last_used): (ImageId, MediaLibrary, DateTime<Utc>) =
-            db::image::recent::create(db.as_ref(), claims.0.user_id, req.id, req.library).await?;
-
-        Ok(CreatedJson(UserRecentImageResponse {
-            id,
-            library,
-            last_used,
-        }))
-    }
-
-    #[api_v2_operation]
-    pub(in super::super) async fn update(
-        db: Data<PgPool>,
-        claims: TokenUser,
-        req: Path<ImageId>,
-    ) -> Result<NoContent, error::UserRecentImage> {
-        db::image::recent::update(db.as_ref(), claims.0.user_id, req.into_inner()).await?;
-
-        Ok(NoContent)
-    }
-
-    #[api_v2_operation]
-    pub(in super::super) async fn delete(
-        db: Data<PgPool>,
-        claims: TokenUser,
-        req: Path<ImageId>,
-    ) -> Result<NoContent, error::UserRecentImage> {
-        db::image::recent::delete(db.as_ref(), claims.0.user_id, req.into_inner()).await?;
-
-        Ok(NoContent)
-    }
-
-    #[api_v2_operation]
-    pub(in super::super) async fn list(
-        db: Data<PgPool>,
-        claims: TokenUser,
-        req: Option<Query<<recent::List as ApiEndpoint>::Req>>,
-    ) -> Result<Json<<recent::List as ApiEndpoint>::Res>, error::UserRecentImage> {
-        // Handle optional limit here.
-        // `None` becomes `limit null` in postgres query. This means no limit.
-        let limit = match req {
-            None => None,
-            Some(query) => {
-                let limit = query.into_inner().limit;
-                Some(limit as i64)
-            }
-        };
-
-        let images: Vec<UserRecentImageResponse> =
-            db::image::recent::list(db.as_ref(), claims.0.user_id, limit).await?;
-
-        Ok(Json(UserRecentImageListResponse { images }))
-    }
-}
+pub mod recent;
+pub mod tag;
+pub mod user;
 
 /// Create an image in the global image library.
 #[api_v2_operation]
