+++ shared/rust/src/api/endpoints/image/user.rs
@@ -1,0 +1,70 @@
+//! Routes for the user image library
+
+use crate::{
+    api::{ApiEndpoint, Method},
+    domain::{
+        image::{
+            user::{UserImageListResponse, UserImageResponse},
+            ImageId,
+        },
+        CreateResponse,
+    },
+    error::EmptyError,
+};
+
+/// List user library images.
+pub struct List;
+
+impl ApiEndpoint for List {
+    type Req = ();
+    type Res = UserImageListResponse;
+    type Err = EmptyError;
+    const PATH: &'static str = "/v1/user/me/image";
+    const METHOD: Method = Method::Get;
+}
+
+/// Get an user library image by ID.
+pub struct Get;
+
+impl ApiEndpoint for Get {
+    type Req = ();
+    type Res = UserImageResponse;
+    type Err = EmptyError;
+    const PATH: &'static str = "/v1/user/me/image/{id}";
+    const METHOD: Method = Method::Get;
+}
+
+/// Create an user library image.
+pub struct Create;
+
+impl ApiEndpoint for Create {
+    type Req = ();
+    type Res = CreateResponse<ImageId>;
+    type Err = EmptyError;
+    const PATH: &'static str = "/v1/user/me/image";
+    const METHOD: Method = Method::Post;
+}
+
+/// Upload an image to the user image library.
+/// Note: can be used to update the raw data associated with the image.
+pub struct Upload;
+
+impl ApiEndpoint for Upload {
+    // raw bytes
+    type Req = ();
+    type Res = ();
+    type Err = EmptyError;
+    const PATH: &'static str = "/v1/user/me/image/{id}/raw";
+    const METHOD: Method = Method::Put;
+}
+
+/// Delete an image from the user library.
+pub struct Delete;
+
+impl ApiEndpoint for Delete {
+    type Req = ();
+    type Res = ();
+    type Err = EmptyError;
+    const PATH: &'static str = "/v1/user/me/image/{id}";
+    const METHOD: Method = Method::Delete;
+}
